+++ tide/core/resources/MissingVirtualKeyboard.qml
@@ -1,0 +1,9 @@
+import QtQuick 2.0
+
+Rectangle {
+    color: "gray"
+    Text {
+        text: "Virtual Keyboard not loaded"
+        anchors.centerIn: parent
+    }
+}